refactor: 랜덤 생성하는 클래스 인자 변경
