--- src/main/java/ladder/domain/RandomBasedBarGenerator.java
@@ -1,12 +1,0 @@
-package ladder.domain;
-
-import java.util.Random;
-
-public class RandomBasedBarGenerator implements BarGenerator {
-    private final Random random = new Random();
-
-    @Override
-    public boolean createBar() {
-        return random.nextBoolean();
-    }
-}
